+++ libraries/AD524X/AD524X.h
@@ -2,7 +2,7 @@
 //
 //    FILE: AD524X.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.4.2
+// VERSION: 0.5.0
 // PURPOSE: I2C digital PotentioMeter AD5241 AD5242
 //    DATE: 2013-10-12
 //     URL: https://github.com/RobTillaart/AD524X
@@ -12,7 +12,7 @@
 #include "Wire.h"
 
 
-#define AD524X_LIB_VERSION    (F("0.4.2"))
+#define AD524X_LIB_VERSION    (F("0.5.0"))
 
 
 #define AD524X_OK             0
@@ -27,9 +27,6 @@ class AD524X
 public:
   AD524X(const uint8_t address, TwoWire *wire = &Wire);
 
-#if defined (ESP8266) || defined(ESP32)
-  bool    begin(uint8_t sda, uint8_t scl);
-#endif
   bool    begin();
   bool    isConnected();
 
@@ -70,13 +67,13 @@ protected:
   uint8_t _O1;
   uint8_t _O2;
 
-  TwoWire*  _wire;
+  TwoWire * _wire;
 };
 
 
 //////////////////////////////////////////////////////////////
 //
-//  DERIVED CLASSES
+//  DERIVED CLASSES  AD5241 AD5242
 //
 class AD5241 : public AD524X
 {
@@ -95,6 +92,30 @@ class AD5242 : public AD524X
 {
 public:
   AD5242(const uint8_t address, TwoWire *wire = &Wire);
+};
+
+
+//////////////////////////////////////////////////////////////
+//
+//  DERIVED CLASSES  AD5280 AD5282
+//
+class AD5280 : public AD524X
+{
+public:
+  AD5280(const uint8_t address, TwoWire *wire = &Wire);
+
+  uint8_t write(const uint8_t value);
+  uint8_t write(const uint8_t value, const uint8_t O1, const uint8_t O2);
+
+  uint8_t write(const uint8_t rdac, const uint8_t value);
+  uint8_t write(const uint8_t rdac, const uint8_t value, const uint8_t O1, const uint8_t O2);
+};
+
+
+class AD5282 : public AD524X
+{
+public:
+  AD5282(const uint8_t address, TwoWire *wire = &Wire);
 };