+++ libraries/AD524X/test/unit_test_001.cpp
@@ -51,6 +51,35 @@ unittest(test_constructors)
   AD5242 AD2(0x2C);
   assertEqual(127, AD2.read(0));
   assertEqual(127, AD2.read(1));
+
+  AD5280 AD3(0x2C);
+  assertEqual(127, AD1.read(0));
+  assertEqual(127, AD1.read(1));
+
+  AD5282 AD4(0x2C);
+  assertEqual(127, AD2.read(0));
+  assertEqual(127, AD2.read(1));
+}
+
+
+unittest(test_pmCount)
+{
+  Wire.begin();
+
+  AD524X ADx(0x2C);  //  AD0 & AD1 == GND
+  assertEqual(2, ADx.pmCount());
+  
+  AD5241 AD1(0x2C);
+  assertEqual(2, ADx.pmCount());
+
+  AD5242 AD2(0x2C);
+  assertEqual(2, ADx.pmCount());
+
+  AD5280 AD3(0x2C);
+  assertEqual(2, ADx.pmCount());
+
+  AD5282 AD4(0x2C);
+  assertEqual(2, ADx.pmCount());
 }